Update openpose_editor.py
I haven't tested this, but I googled an alternative to the command you used. hopefully it works? A guy called grandfromages provided a fix for the issue "sending to controlnet" by changing this to "cn_max = 2", because for some reason its 0 and means we can't send to controlnet. But I tried to use a less hacky workaround, but hopefully you can use it to work out the proper command needed?
https://github.com/nonnonstop/sd-webui-3d-open-pose-editor/issues/96